Make project config part of EnsimeLauncher's state (#293)
* Make project config part of EnsimeLauncher's state

This seems logical, and the refactoring reduces the parameters that
callers have to keep track of and pass around through various methods.
It also eliminates wastefully parsing the .ensime file twice, once in
the no_classpath_file function and again in launch.

* DRY location of bootstrap classpath file in Launcher

Again, make this an attribute of the launcher class reduces a lot of
redundant code and function arguments.

* Make constants constant

* Prefix all user feedback messages with "[ensime]"

* Inform the user when we can't run their java executable

This improves a poor user experience when the plugin fails to load and
the reason is unclear to the user. See #280.

* Don’t reinvent string.Template

* Encapsulate legacy bootstrap migration within EnsimeLauncher

Triggering a side effect from an __init__ is a bit smelly but it’s
certainly no worse than triggering one when a module is loaded.