Factor all Vim usage from EnsimeClient into an Editor object
Get rid of the `commands` dictionary of Vim commands/expressions, giving
them a real function API instead of a stringly-typed one.

This is also a push toward unit-testability, reducing complex
dependencies of objects so they can feasibly be injected with mocks.

This is a first pass only, the Editor object API can definitely be
improved but most functions are ported intact at this time to make
review easier and defer design discussions. Also, a few uses of the
`vim` object remain on EnsimeClient -- these are limited to the
disable_plugin functionality and will be addressed in #294, it is too
complicated to include in the scope of this change.

Closes #321.