Put the Syntastic checker in its standard supported place
Syntastic automatically scans the runtimepath for syntax_checkers
directories to accommodate plugins:

https://github.com/scrooloose/syntastic/wiki/Syntax-Checker-Guide#external

The non-standard location was specifically requested in #184, but I
don’t agree with the reasoning or the extra code that is required to
work around it. I’m not sure Jorge understood that Syntastic has a
supported convention.

This reverts this commit, but it was squashed away before merge so the
object isn’t actually reachable from most clones:

https://github.com/ensime/ensime-vim/commit/ed0e941e25ccda3e1367130a9fb23610631d82ac