Begin to clean up the Ensime class interface
We’ve had some recent discussion about what the responsibily of this
class should be, and based on that it has some methods that don’t fit
or at least should not be public. This is a first pass at cleaning up.

Some consideration of the class’s role was hashed out on #328. These
changes have eventual relevance to #243, #294, #325.